requirements.txt
# LangChain 相关
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-mcp-adapters>=0.1.0

# 配置管理
python-decouple>=3.8

# 异步支持
aiohttp>=3.8.0

# 工具类
pydantic>=2.0.0
typing-extensions>=4.0.0
httpx[socks]